Add support for Apple Silicon
This commit adds support for Apple Silicon. Primarily, this means the addition of an architecture check and associated #define to remove x86-specific code from the build (base9/sleef directory) and to provide Apple Silicon-compatible versions of the margEvolveNoBinaries function (which is SIMD-optimized).

Some minor additional changes were needed to compiler flags to remove unsupported flags when building on AS.

A new single-core implementation of the parallel reference version of margEvolveNoBinaries has been provided as the default implementation. The parallel implementation is still available behind a BUILD_PARALLEL flag, but no facility is provided to enable it at this time.

As a "while I'm in there", CMake "minimum version" specifications have been updated to specify the newest version as an upper bound. This removes "may be deprecated" warnings from all BASE-9 code (though not from the YAML-cpp build)."